% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ADCingr.R
\name{calculate_ADCingr}
\alias{calculate_ADCingr}
\title{Apparent Digestibility Coefficient of a feed ingredient (ADCingr)}
\usage{
calculate_ADCingr(ADC_test, ADC_ref, nutr_ref, nutr_ingr, incl_ingr)
}
\arguments{
\item{ADC_test}{a numeric value in the interval [0,1] that represents the
Apparent Digestibility Coefficient (ADC) of the diet that contains the
ingredient to be tested.}

\item{ADC_ref}{a numeric value in the interval [0,1] that represents the
Apparent Digestibility Coefficient (ADC) of the reference diet without the
ingredient to be tested.}

\item{nutr_ref}{a numeric value in the interval [0,1] that represents the
inclusion rate of the nutrient in the reference diet.}

\item{nutr_ingr}{a numeric value in the interval [0,1] that represents the
inclusion rate of the nutrient in the test diet for which the Apparent
Digestibility Coefficient (ADC) of the nutrient in the ingredient will be
calculated.}

\item{incl_ingr}{a numeric value in the interval [0,1] that represents the
inclusion rate of the ingredient in the test diet for which the Apparent
Digestibility Coefficient (ADC) of the nutrient in an ingredient will be
calculated.}
}
\value{
returns a single numeric value which is the relative ADC for the
diet. If the value is not within the interval [0,1], an additional warning
is returned.
}
\description{
Function to calculate the Apparent Digestibility Coefficient (ADC) of a feed
ingredient that is part of a compound diet.
}
\note{
The calculation of the ADCingr is based on the equation proposed by Bureau
and Hua (2006).
}
\references{
Bureau, D. P., & Hua, K. (2006): Letter to the Editor of
Aquaculture. Aquaculture, 252, p.103–105.
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
Anıl Axel Tellbüscher
}
