% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_NER.R
\name{calculate_NER}
\alias{calculate_NER}
\title{NER}
\usage{
calculate_NER(
  initWeight = NULL,
  finalWeight = NULL,
  weightGain = NULL,
  nutrFed = NULL
)
}
\arguments{
\item{initWeight}{a numeric value for the initial weight (either average
weight of the individuals or the total biomass) of the livestock at the
beginning of the feeding trial.}

\item{finalWeight}{a numeric value for the final weight (either average
weight of the individuals or the total biomass) of the livestock at the end
of the feeding trial.}

\item{weightGain}{optional; can be provided instead of the initial and final
weight.}

\item{nutrFed}{a numeric value for the amount of nutrient fed during the
feeding trial.}
}
\value{
returns a single numeric value
}
\description{
Function to calculate the Nutrient Efficiency Ratio (NER)
}
\details{
It should be noted that all input values must be provided in the same unit.
}
\note{
%% ~~further notes~~
}
\examples{
# Calculate the NER using the initial and final weight
calculate_NER(initWeight = 1, finalWeight = 10, nutrFed = 12)

# Calculate the NER using the weight gain instead
calculate_NER(weightGain = 9, nutrFed = 12)

}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
\code{\link[Aquaculture Formulary]{https://blog.tellbuescher.online/formulary}}
}
\author{
Anıl Axel Tellbüscher
}
