% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{digestingr}
\alias{digestingr}
\title{Digestibility of a feed ingredient}
\format{
A tibble with one row and six columns
\describe{
 \item{diet_reference}{character; reference feed identifier.}
 \item{ingredient}{character; feed ingredient identifier.}
 \item{adc_reference}{numeric; apparent digestibility coefficient of the 
 reference feed.}
 \item{adc_test}{numeric; apparent digestibility coefficient of the test 
 feed.}
 \item{dm_ingr}{numeric; mass fraction of dry matter in the feed ingredient. 
 Data reported as relative value between 0 and 1.}
 \item{N_ingr}{numeric; mass fraction of nitrogen in the feed ingredient 
 in g/kg.}
}
}
\description{
This dataset contains example data for the determination of the apparent 
digestibility of feed ingredients by using a reference diet and replacing 
a part (usually 30%) of it by the feed ingredient to be evaluated.
}
\examples{
digestingr

}
\keyword{dataset}
