% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{digestnut}
\alias{digestnut}
\title{Digestibility of feed nutrients}
\format{
A tibble with two rows and six columns:
\describe{
 \item{diet}{character; feed identifier.}
 \item{dm}{numeric; mass fraction of dry matter in the feed. Data reported 
 as value between 0 and 1, corresponding to g/g.}
 \item{N_feed}{numeric; mass fraction of nitrogen in feed. Data reported as 
 value between 0 and 1, corresponding to g/g.}
 \item{std_feed}{numeric; mass fraction of digestibility marker in feed.
 Data reported as value between 0 and 1, corresponding to g/g.}
 \item{N_feces}{numeric; mass fraction of nitrogen in feces in g/kg.}
 \item{std_feces}{numeric; mass fraction of digestibility marker in feces. 
 Data reported as value between 0 and 1, corresponding to g/g.}
}
}
\description{
This dataset contains example data for the calculation of apparent 
digestibility coefficients for the digestibility of nitrogen in the dry 
matter fraction of an aquaculture feed.
}
\examples{
digestnut

}
\keyword{dataset}
