% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr.R
\name{fr}
\alias{fr}
\title{Feeding rate (FR)}
\usage{
fr(m_feed = NULL, duration = NULL)
}
\arguments{
\item{m_feed}{numeric; total amount of feed given in grams over the period of 
interest.}

\item{duration}{numeric; time period}
}
\value{
numeric
}
\description{
Feed administered per time unit. The feeding rate usually refers to the 
amount of feed given, but not to the true feed intake.
}
\author{
Anıl Axel Tellbüscher
}
