% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rearing}
\alias{rearing}
\title{Rearing data from the first polyculture experiment}
\format{
A data frame containing 24 rows and four columns:
\describe{
   \item{tank}{character; unique identifier for each fish tank.}
   \item{feed_g}{numeric; total feed input in gram over the duration of the 
   experiment.}
   \item{temp_C}{numeric; average water temperature in degrees Celsius.}
   \item{duration}{numeric; total duration of the experiment in days.}
}
}
\description{
A dataset containing feed quantities fed per tank and the average water 
temperature over all tanks stocked with pikeperch oder sturgeon. The 
variables are as follows:
}
\examples{
rearing

}
\keyword{datasets}
