% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retention.R
\name{retention}
\alias{retention}
\title{Nutrient retention (NR)}
\usage{
retention(m_start, nut_start, m_end, nut_end)
}
\arguments{
\item{m_start}{numeric; value that is providing the initial weight in grams}

\item{nut_start}{numeric; value between 0 and 1 that provides the initial
inclusion rate of the target nutrient in the tissue of the experimental animal.}

\item{m_end}{numeric; value that is providing the final weight in grams}

\item{nut_end}{numeric; value between 0 and 1 that provides the final inclusion
rate of the target nutrient in the tissue of the experimental animal.}
}
\value{
a numeric value
}
\description{
Function to calculate the nutrient retention
}
\author{
Anıl Axel Tellbüscher
}
