% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treatments}
\alias{treatments}
\title{Treatments}
\format{
A tibble with nine rows and two columns:
\describe{
 \item{tank}{factor; rearing tank identifier.}
 \item{treatment}{factor; treatment identifier.}
}
}
\description{
This dataset contains the rearing tank and respective treatment assigned to 
it. It represents the metadata from a feeding trial with Atlantic salmon 
(Salmo salar). The related dataset was published by Liland et al. (2024) 
and is from the first out of two trials ("Trial A").
}
\examples{
treatments

}
\references{
Liland, N., Rønnestad, I., Azevedo, M., Lai, F., Oulie, F., 
Conceição, L., Soares, F. (2024): Dataset on the performance of Atlantic 
salmon (Salmo salar) reared at different dissolved oxygen levels under 
experimental conditions. Data in Brief 57, 110983. 
https://doi.org/10.1016/j.dib.2024.110983
}
\keyword{datasets}
