% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adc_dm.R
\name{adc_dm}
\alias{adc_dm}
\title{ADC of DM}
\usage{
adc_dm(es_diet, es_feces)
}
\arguments{
\item{es_diet}{a numeric value, resembling the inclusion rate of external
standard in the experimental diet given to the livestock.}

\item{es_feces}{a numeric value, resembling the inclusion rate of external
standard in the feces recovered during the digestibility trial.}
}
\value{
returns a single numeric value which is the relative ADC for the
dry matter content of the diet. If the value is not within the interval [0,1], 
an additional warning is returned.
}
\description{
Function to calculate the Apparent Digestibility Coefficient (ADC) of the dry 
matter fraction of a compound diet.
}
\references{
Bureau, D. P., & Hua, K. (2006): Letter to the Editor of
Aquaculture. Aquaculture, 252, p.103–105.
}
\author{
Anıl Axel Tellbüscher
}
