% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adc_nut.R
\name{adc_nut}
\alias{adc_nut}
\title{ADC of a nutrient (ADCnut)}
\usage{
adc_nut(es_diet, es_feces, nut_diet, nut_feces)
}
\arguments{
\item{es_diet}{numeric; value resembling the inclusion rate of external
standard in the experimental diet given to the livestock.}

\item{es_feces}{numeric; value resembling the inclusion rate of external
standard in the feces recovered during the digestibility trial.}

\item{nut_diet}{numeric; value representing the inclusion rate of the 
target nutrient in the diet.}

\item{nut_feces}{numeric; value representing the inclusion rate of the 
target nutrient in the feces.}
}
\value{
returns a single numeric value which is the relative ADC for a
single nutrient in the diet. If the value is not within the interval [0,1],
an additional warning is returned.
}
\description{
Function to calculate the Apparent Digestibility Coefficient of a nutrient
in the dry matter fraction of a compound diet.
}
\references{
Bureau, D. P., & Hua, K. (2006): Letter to the Editor of
Aquaculture. Aquaculture, 252, p.103–105.
}
\author{
Anıl Axel Tellbüscher
}
