% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afr.R
\name{afr}
\alias{afr}
\title{Absolute feeding rate (AFR)}
\usage{
afr(m_feed = NULL, duration = NULL)
}
\arguments{
\item{m_feed}{numeric; total amount of feed provided in grams over the 
time period of interest.}

\item{duration}{numeric; time period over which the animals are fed}
}
\value{
numeric; returns a vector if multiple values are provided.
}
\description{
Feed administered per unit of time. The absolute feeding rate usually refers 
to the amount of feed given, but not to the true feed intake.
}
\examples{
# 100 g fed over 10 d
afr(m_feed = 100, duration = 10)

# 354 g, 846 g, and 1222 g fed over 83 d
afr(m_feed = c(354, 846, 1222), duration = 83)

}
\author{
Anıl Axel Tellbüscher
}
