% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag.R
\name{ag}
\alias{ag}
\title{Absolute Growth (AG)}
\usage{
ag(m_start, m_nend)
}
\arguments{
\item{m_start}{numeric; value providing the initial weight in grams.}

\item{m_end}{numeric; value providing the final weight in grams.}
}
\value{
returns a numeric value that is the AG.
}
\description{
Calculates the Absolute Growth
}
\details{
A function that calculates the Absolute Growth (AG), also denoted as
Absolute Weight Gain (AWG) based on the Initial Body weight (IBW;
initWeight) in gram (g) and the Final Body Weight (FBW; finalWeight)
in gram (g).
}
\note{
The AG, also termed Absolute Weight Gain (AWG), is the weight that was
gained within the duration of a growth trial. It does by default not
distinguish between weight gain in form of dry matter or water.

AG = FBW - IBW

The IBW is subtracted from the FBW.
}
\references{
Lugert, V., Thaller, G., Tetens, J., Schulz, C., & Krieter, J.
(2016): A review on fish growth calculation: multiple functions in fish
production and their specific application. Reviews in Aquaculture, 8,
p.30–42.
}
\seealso{
\code{\link[Aquaculture Formulary]{https://blog.tellbuescher.online/formulary}}
}
\author{
Anıl Axel Tellbüscher
}
