% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agr.R
\name{agr}
\alias{agr}
\title{Absolute Growth Rate (AGR)}
\usage{
agr(m_start, m_end, duration)
}
\arguments{
\item{m_start}{a numeric value providing the initial weight in grams.}

\item{m_end}{a numeric value providing the final weight in grams.}

\item{duration}{a numeric value describing the duration of the growth experiment.}
}
\value{
returns a numeric value that is the AGR.
}
\description{
A function that calculates the Absolute Growth Rate (AGR), also denoted as 
Weight Gain (WG) when applied to weight data.
}
\examples{
data(weight2)

dplyr::mutate(weight2, AGR = agr(W_start, W_end, duration = 84))



}
\references{
Lugert, V., Thaller, G., Tetens, J., Schulz, C., & Krieter, J.
(2016): A review on fish growth calculation: multiple functions in fish
production and their specific application. Reviews in Aquaculture, 8,
p.30–42.
}
\author{
Anıl Axel Tellbüscher
}
