% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{digestdm}
\alias{digestdm}
\title{Digestibility of dry matter}
\format{
A tibble with two rows and four columns:
\describe{
 \item{diet}{character; feed identifier.}
 \item{dm}{numeric; mass fraction of dry matter in the feed.}
 \item{std_feed}{numeric; mass fraction of digestibility marker in feed.}
 \item{std_feces}{numeric; mass fraction of digestibility marker in feces.}
}
}
\description{
This dataset contains example data to calculate the apparent digestibility
of two diets that differ in their dry matter mass fraction and in the mass 
fraction of digestibility markers in the feces.
}
\examples{
digestdm

}
\keyword{datasets}
