% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fce.R
\name{fce}
\alias{fce}
\title{Feed conversion efficiency (FCE)}
\usage{
fce(m_start, m_end, fi, dm = 1)
}
\arguments{
\item{m_start}{numeric; value that is providing the initial weight in
grams.}

\item{m_end}{numeric; value that is providing the final weight in
grams.}

\item{fi}{numeric; value providing the total feed intake in grams
during the experiment.}

\item{dm}{numeric; value within the interval of (0:1), indicating the dry 
matter content of the feed.}
}
\value{
a numeric value
}
\description{
A function that calculates the feed conversion efficiency (FCE), which is the 
inverse of the feed conversion ratio
}
\author{
Anıl Axel Tellbüscher
}
