% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr.R
\name{fcr}
\alias{fcr}
\title{Feed Conversion Ratio (FCR)}
\usage{
fcr(m_start, m_end, fi, dm)
}
\arguments{
\item{m_start}{numeric; value that is providing the initial weight in
grams.}

\item{m_end}{numeric; value that is providing the final weight in
grams.}

\item{fi}{numeric; value providing the total feed intake in grams
during the experiment.}

\item{dm}{numeric; value indicating the dry matter content of the
feed. The value must be within the interval of (0:1).}
}
\value{
returns a numeric value that is the FCR.
}
\description{
A function that calculates the Feed Conversion Ratio (FCR) based on
the Initial Body weight (IBW; initWeight) in gram (g), the Final Body
Weight (FBW; finalWeight) in gram (g), the total Feed Intake (FI; feedIntake)
in gram and the dry matter (DM; dryMatter) content of the feed in percent.
(duration) in days (d).
}
\note{
The Feed Conversion Ratio (FCR) describes the amount of feed on dry matter
(DM) basis that is required to gain 1 kg of body weight on wet weight basis.

FCR = (FI * DM) / (FBW - IBW)

The FI is multiplied by the DM of the feed to calculate the true FI.
This is then divided by the weight increase during the experimental
period to calculate the amount of feed consumed per unit of weight
increase.
}
\references{
Lugert, V., Thaller, G., Tetens, J., Schulz, C., & Krieter, J.
(2016): A review on fish growth calculation: multiple functions in fish
production and their specific application. Reviews in Aquaculture, 8,
p.30–42.
}
\seealso{
\code{\link[Aquaculture Formulary]{https://blog.tellbuescher.online/formulary}}
}
\author{
Anıl Axel Tellbüscher
}
