% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feedcomp}
\alias{feedcomp}
\title{Feed composition}
\format{
A data frame with one row and 25 columns:
\describe{
 \item{diet}{character; feed name}
 \item{dry_matter}{numeric; content of feed in percent}
 \item{crude_protein}{numeric; content of feed in percent as fed}
 \item{crude_lipids}{numeric; content of feed in percent as fed}
 \item{ash}{numeric; content of feed in percent as fed}
 \item{gross_energy}{numeric; content of feed in MJ per kg as fed}
 \item{phosphorus}{numeric; content of feed in percent as fed}
 \item{arginine}{numeric; content of feed in percent as fed}
 \item{histidine}{numeric; content of feed in percent as fed}
 \item{isoleucine}{numeric; content of feed in percent as fed}
 \item{leucine}{numeric; content of feed in percent as fed} 
 \item{lysine}{numeric; content of feed in percent as fed}
 \item{threonine}{numeric; content of feed in percent as fed}
 \item{tryptophan}{numeric; content of feed in percent as fed}
 \item{valine}{numeric; content of feed in percent as fed}
 \item{methionine}{numeric; content of feed in percent as fed}
 \item{cysteine}{numeric; content of feed in percent as fed}
 \item{phenylalanine}{numeric; content of feed in percent as fed}
 \item{tyrosine}{numeric; content of feed in percent as fed}
 \item{aspartic_acid}{numeric; content of feed in percent as fed}
 \item{glutamic_acid}{numeric; content of feed in percent as fed}
 \item{alanine}{numeric; content of feed in percent as fed}
 \item{glycine}{numeric; content of feed in percent as fed}
 \item{proline}{numeric; content of feed in percent as fed}
 \item{serine}{numeric; content of feed in percent as fed}
}
}
\description{
This dataset contains compositional data of Skretting Protec, a commercial
fish feed. The data comprises the proximate composition, phosphorus, and
some essential amino acids.
}
\examples{
feedcomp

}
\references{
Liland, N., Rønnestad, I., Azevedo, M., Lai, F., Oulie, F., 
Conceição, L., Soares, F. (2024): Dataset on the performance of Atlantic 
salmon (Salmo salar) reared at different dissolved oxygen levels under 
experimental conditions. Data in Brief 57, 110983. 
https://doi.org/10.1016/j.dib.2024.110983
}
\keyword{datasets}
