% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbw.R
\name{gbw}
\alias{gbw}
\title{Geometric Mean Bodyweight (GMBW)}
\usage{
gbw(ibw, fbw)
}
\arguments{
\item{ibw}{numeric; initial bodyweight weight in grams.}

\item{fbw}{numeric; final bodyweight in grams.}
}
\value{
numeric value that is the geometric mean bodyweight.
}
\description{
A function that calculates the geometric mean of the initial and final 
bodyweight.
}
\details{
While the arithmetic mean assumes a linear relationship between the averaged
numbers, the geometric mean accounts for the non-linear and potentially variable 
nature of animal growth.
}
\examples{
data(weight2)

dplyr::mutate(weight2, GMBW = gbw(W_start, W_end))



}
\author{
Anıl Axel Tellbüscher
}
