% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbw.R
\name{gbw}
\alias{gbw}
\title{Geometric Bodyweight (GBW)}
\usage{
gbw(m_start, m_end)
}
\arguments{
\item{m_start}{a numeric value that is providing the initial weight in
grams.}

\item{m_end}{a numeric value that is providing the final weight in
grams.}
}
\value{
returns a numeric value that is the geometric bodyweight.
}
\description{
A function that calculates the geometric mean bodyweight.
}
\examples{
data(weight2)

dplyr::mutate(weight2, GBW = gbw(W_start, W_end))



}
\author{
Anıl Axel Tellbüscher
}
