% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbw.R
\name{mbw}
\alias{mbw}
\title{Metabolic bodyweight (MBW)}
\usage{
mbw(m_start, m_end, gbw = NULL)
}
\arguments{
\item{m_start}{a numeric value that is providing the initial weight in
grams.}

\item{m_end}{a numeric value that is providing the final weight in
grams.}

\item{gbw}{a numeric value providing the geometric bodyweight in grams.}
}
\value{
returns a numeric value that is the metabolic bodyweight.
}
\description{
The metabolic bodyweight is the rate of energy expenditure in dependence of
the bodyweight of an organism.
}
\note{
The calculation is based on Lupatsch et al. (2003), with an exponent of 0.8.
}
\examples{
data(weight2)

dplyr::mutate(weight2, MBW = mbw(W_start, W_end))



}
\references{
Lupatsch, I. et al. (2003): Comparison of energy and protein
efficiency among three fish species gilthead sea bream (Sparus aurata),
European sea bass (Dicentrarchus labrax) and white grouper (Epinephelus
aeneus): energy expenditure for protein and lipid deposition. Aquaculture
225, p.175-189.
}
\author{
Anıl Axel Tellbüscher
}
