% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearing.R
\docType{data}
\name{rearing}
\alias{rearing}
\title{Rearing data from the first polyculture experiment}
\format{
A data frame with 24 rows and 4 variables.
}
\usage{
data(rearing)
}
\description{
A dataset containing feed quantities fed per tank and the average water 
temperature over all tanks stocked with pikeperch oder sturgeon. The 
variables are as follows:
}
\details{
\itemize{
   \item tank. unique identifier for each fish tank.
   \item feed_g. feed input in gram over the duration of the experiment.
   \item temp_C. water temperature in degrees Celsius.
   \item duration. duration of the experiment.
}
}
\keyword{datasets}
