% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfr.R
\name{rfr}
\alias{rfr}
\title{Relative Feeding Rate (FR)}
\usage{
rfr(m_feed = NULL, m_bio = NULL, duration = NULL)
}
\arguments{
\item{m_feed}{numeric; total amount of feed given in grams over the period of 
interest.}

\item{m_bio}{numeric}

\item{duration}{numeric; time period}
}
\value{
numeric
}
\description{
Feed administered per time unit and biomass. The feeding rate usually 
refers to the amount of feed given, but not to the true feed intake.
}
\author{
Anıl Axel Tellbüscher
}
