% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg.R
\name{rg}
\alias{rg}
\title{Relative Growth (RG)}
\usage{
rg(ibw, fbw)
}
\arguments{
\item{ibw}{numeric; value providing the initial weight in grams.}

\item{fbw}{numeric; value providing the final weight in grams.}
}
\value{
returns a numeric value that is the Relative Growth. Multiply by 100
for conversion into a percentage.
}
\description{
A function that calculates the Relative Growth (RG) based on the
Initial Body weight (IBW; ibw) and the Final Body Weight (FBW; fbw)
in grams (g).
}
\examples{
data(weight2)
dplyr::mutate(weight2, RG = rg(ibw_g, fbw_g))

}
\author{
Anıl Axel Tellbüscher
}
