% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgr.R
\name{rgr}
\alias{rgr}
\title{Relative Growth Rate (RGR)}
\usage{
rgr(m_start, m_end, duration)
}
\arguments{
\item{m_start}{numeric; provides the initial weight in gram.}

\item{m_end}{numeric; provides the final weight in gram.}

\item{duration}{numeric; duration of the growth experiment.}
}
\value{
either a single numeric value or vector holding the calculated RGR
values. Multiply by 100 for conversion into percentage.
}
\description{
A function that calculates the Relative Growth Rate (RGR), which is the 
relative weight increase per time unit.
}
\examples{
data(weight2)
dplyr::mutate(weight2, RGR = rgr(ibw_g, fbw_g, duration = 84))

}
\references{
Lugert, V., Thaller, G., Tetens, J., Schulz, C., & Krieter, J.
(2016): A review on fish growth calculation: multiple functions in fish
production and their specific application. Reviews in Aquaculture, 8,
p.30–42.
}
\author{
Anıl Axel Tellbüscher
}
