% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgr.R
\name{rgr}
\alias{rgr}
\title{Relative Growth Rate (RGR)}
\usage{
rgr(m_start, m_end, duration)
}
\arguments{
\item{m_start}{a numeric value that is providing the initial weight in
grams.}

\item{m_end}{a numeric value that is providing the final weight in
grams.}

\item{duration}{a numeric value describing the duration of the growth experiment.}
}
\value{
returns a numeric value that is the RGR.
}
\description{
A function that calculates the Relative Growth Rate (RGR).
}
\references{
Lugert, V., Thaller, G., Tetens, J., Schulz, C., & Krieter, J.
(2016): A review on fish growth calculation: multiple functions in fish
production and their specific application. Reviews in Aquaculture, 8,
p.30–42.
}
\author{
Anıl Axel Tellbüscher
}
