% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\docType{data}
\name{weight}
\alias{weight}
\title{Fish biomass and bodyweights from the first polyculture experiment}
\format{
A data frame with 53940 rows and 10 variables
}
\usage{
data(weight)
}
\description{
A dataset containing fish weights
}
\keyword{datasets}
