% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight2.R
\docType{data}
\name{weight2}
\alias{weight2}
\title{Individual fish bodyweights during a polyculture experiment}
\format{
A tibble with 396 rows and four columns:
\describe{
 \item{species}{character; species identifier.}
 \item{culture_sys}{character; culture system identifier.}
 \item{ibw_g}{numeric; initial bodyweight in gram.}
 \item{fbw_g}{numeric; final bodyweight in gram.}
}
}
\description{
A dataset holding weight data of tagged individuals of two different fish 
species, pikeperch (Sander lucioperca) and largemouth bass (Micropterus 
salmoides), at the beginning and the end of a polyculture experiment for 
aquaculture purposes.
}
\examples{
weight2

}
\references{
Pěnka T., Malinovskyi O., Křišťan J., Imentai A., Policar T. 
(2021): Effect of density and mixed culture with pikeperch (Sander 
lucioperca) on effectivity of largemouth bass (Micropterus salmoides) 
intensive culture. Czech Journal of Animal Science, 66: 428–440.
}
\keyword{datasets}
