% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight2.R
\docType{data}
\name{weight2}
\alias{weight2}
\title{Fish biomass and bodyweights from the second polyculture experiment}
\format{
A data frame with 53940 rows and 10 variables
}
\usage{
data(weight2)
}
\description{
A dataset containing fish weights
}
\keyword{datasets}
